+++ src/model/Preferences.js
@@ -1,5 +1,7 @@
 // @flow
 
+import type { DesiredPart } from "./DesiredPart";
+
 export type EnabledAlts = Array<string>;
 
 export function getEnabledAlts(): EnabledAlts {
@@ -9,6 +11,16 @@ export function getEnabledAlts(): EnabledAlts {
 
 export function setEnabledAlts(newEnabledAlts: EnabledAlts): void {
   saveEnabledAlts(newEnabledAlts);
+}
+
+export function loadDesiredParts(): Array<DesiredPart> {
+  const json = localStorage.getItem("desiredParts") || "[]";
+  return JSON.parse(json);
+}
+
+export function saveDesiredParts(newDesiredParts: Array<DesiredPart>): void {
+  const json = JSON.stringify(newDesiredParts);
+  localStorage.setItem("desiredParts", json);
 }
 
 function loadEnabledAlts() {